# api/swagger.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/conversion/convert/tsv":{"post":{"tags":["CONVERSION"],"summary":"Upload Tsv File","description":"Upload a TSV file for conversion.","operationId":"upload_tsv_file_api_v1_conversion_convert_tsv_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_tsv_file_api_v1_conversion_convert_tsv_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response 200 Upload Tsv File Api V1 Conversion Convert Tsv Post"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidFileFormatResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalUnhandledErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/conversion/convert/csv":{"post":{"tags":["CONVERSION"],"summary":"Upload Csv File","description":"Upload a CSV file for conversion.","operationId":"upload_csv_file_api_v1_conversion_convert_csv_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_csv_file_api_v1_conversion_convert_csv_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response 200 Upload Csv File Api V1 Conversion Convert Csv Post"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidFileFormatResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalUnhandledErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_csv_file_api_v1_conversion_convert_csv_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_csv_file_api_v1_conversion_convert_csv_post"},"Body_upload_tsv_file_api_v1_conversion_convert_tsv_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_tsv_file_api_v1_conversion_convert_tsv_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InternalUnhandledError":{"properties":{"message":{"type":"string","title":"Message","default":"Got internal error. Investigation is recommended."},"code":{"type":"integer","title":"Code","default":500}},"type":"object","required":["message","code"],"title":"InternalUnhandledError"},"InternalUnhandledErrorResponse":{"properties":{"error":{"allOf":[{"$ref":"#/components/schemas/InternalUnhandledError"}],"default":{"code":500,"message":"Got internal error. Investigation is recommended."}}},"type":"object","required":["error"],"title":"InternalUnhandledErrorResponse"},"InvalidFileFormatError":{"properties":{"message":{"type":"string","title":"Message","default":"Invalid file format."},"code":{"type":"integer","title":"Code","default":400}},"type":"object","required":["message","code"],"title":"InvalidFileFormatError"},"InvalidFileFormatResponse":{"properties":{"error":{"allOf":[{"$ref":"#/components/schemas/InvalidFileFormatError"}],"default":{"code":400,"message":"Invalid file format."}}},"type":"object","required":["error"],"title":"InvalidFileFormatResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}